+++ CHANGES.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [HEAD] - Unreleased
+
+## [0.5.0] - 2016-09-08
 ### Added
 - Added `Grove.LED`.
 - Added `Grove.Potentiometer`.
@@ -28,7 +30,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial release with support for `Grove.Button`.
 
-[HEAD]:  https://github.com/bendiken/nerves_grove/compare/0.4.0...HEAD
+[HEAD]:  https://github.com/bendiken/nerves_grove/compare/0.5.0...HEAD
+[0.5.0]: https://github.com/bendiken/nerves_grove/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/bendiken/nerves_grove/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/bendiken/nerves_grove/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/bendiken/nerves_grove/compare/0.1.0...0.2.0